+++ database/migrations/2019_03_25_034726_create_api_tokens_table.php
@@ -14,8 +14,8 @@ class CreateApiTokensTable extends Migration
     public function up()
     {
         Schema::create('api_tokens', function (Blueprint $table) {
-            $table->increments('id');
-            $table->unsignedInteger('user_id');
+            $table->bigIncrements('id');
+            $table->unsignedBigInteger('user_id');
 
             $table->string('token')->unique();
             $table->boolean('enabled')->default(true);
